Create graphql functions
